# Speedup classification by avoiding to load each push information