--- a/tests/conftest.py
+++ b/tests/conftest.py
@@ -109,7 +109,7 @@ def create_push(monkeypatch, responses):
         push._id = push_id
         push_rev_to_id[rev] = push_id
         push.backedoutby = None
-        push.bugs = {push_id}
+        push._bugs = {push_id}
         push.tasks = []
         push._revs = [push.rev]
         push.is_manifest_level = False