The classification algorithm uses the method `Push.bustage_fixed_by` that trigger an HGMO query for each push, just to get their related Bugzilla bug ID.

This patch uses the `full=1` parameter on HGMO json-pushes method in order to get the description of each revision, and thus their Bugzilla bug ID.

Here is a benchmark running on a random recent autoland revision (command `time mozci push classify-eval --rev=808ac93bdd9847272087ba981173bc877cff0425 --recalculate`)

|            | master    | parse-bug-description |
|------------|-----------|-----------------------|
| cold cache | 1m35,881s | 0m8,662s              |
| warm cache | 0m15,875s | 0m6,002s              |

On another revision which triggers good `time mozci push classify-eval --rev=9c1b66dcaac38344a227cc6cdf3e29272b487c98 --recalculate`, the results are evern better:
- cold cache on master is 2m38,573s
- warm cache on this branch is 0m3,063s

The results stay the same on both branches