@marco-c 2022-01-19
It might be worth making the hgmo properties try to get their data from json-pushes first, so we're going to less frequently hit json-automationrelevance. We can file a follow-up for that if you prefer (it's a minor perf improvement compared to this PR).