@marco-c 2022-01-19
Here or as a follow-up, we can optimize `make_push_objects` to also set the bugs property of the push, given that `push_revisions` is returning the desc now.